fix: signup page flow
Signup page had a few edge cases of callbacks there were not defined. This is fixed now

WT-617 #done
