+++ packages/extension/.env
@@ -1,10 +1,7 @@
-NEXT_PUBLIC_API_URL=http://localhost:4000
+NEXT_PUBLIC_API_URL=http://api.apps.local.app.garden
 #NEXT_PUBLIC_API_URL=https://api.daily.dev
-NEXT_PUBLIC_SUBS_URL=ws://localhost:5000/graphql
-NEXT_PUBLIC_DOMAIN=localhost
-NEXT_PUBLIC_GA=UA-109059578-2
-NEXT_PUBLIC_AMPLITUDE=46a18c8e538f37fbb74150e458b5c574
-NEXT_PUBLIC_FLAGSMITH=LZp89Rk56sytKH9AeHyCtF
+NEXT_PUBLIC_SUBS_URL=ws://api.apps.local.app.garden/graphql
+NEXT_PUBLIC_DOMAIN=apps.local.app.garden
 
-NEXT_PUBLIC_AUTH_URL=http://127.0.0.1:4433
-NEXT_PUBLIC_WEBAPP_URL=http://localhost:5002/
+NEXT_PUBLIC_AUTH_URL=http://sso.apps.local.app.garden
+NEXT_PUBLIC_WEBAPP_URL=http://apps.local.app.garden/
